LPS-105024 Organize directory
