+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-spi/bnd.bnd
@@ -3,4 +3,6 @@ Bundle-SymbolicName: com.liferay.portal.security.crypto.generator.spi
 Bundle-Version: 1.0.0
 Export-Package:\
 	com.liferay.portal.security.crypto.generator.spi,\
-	com.liferay.portal.security.crypto.generator.spi.factory
+	com.liferay.portal.security.crypto.generator.spi.hashing,\
+	com.liferay.portal.security.crypto.generator.spi.hashing.factory,\
+	com.liferay.portal.security.crypto.generator.spi.hashing.salt
