+++ modules/apps/portal-security-crypto/portal-security-crypto-generator-spi/src/main/resources/com/liferay/portal/security/crypto/generator/spi/hashing/salt/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
